requirements.txt
numpy
pandas
opencv-python
torch
torchvision
timm
transformers
Pillow
ultralytics
pytest
